requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sentence-transformers==3.3.1
chromadb==0.5.23
sqlglot==26.11.1
graphql-core==3.2.5
httpx==0.28.1
jinja2==3.1.5
python-multipart==0.0.20
